+++ simple_blocks/src/renderhelp.h
@@ -53,7 +53,7 @@ loop(GLFWwindow *, BlockRenderer *);
 /// \brief Get the total and avail memory. If no pointer to avail is provided
 /// then just total is returned.
 void
-queryGPUMemory(int *total, int *avail = nullptr);
+queryGPUMemory(int64_t *total, int64_t *avail = nullptr);
 
 //void
 //setCameraPosPreset(unsigned int cameraPos);
